Several updates to interface

+++ CodeAthon/codeathon/admin/routes.py
@@ -13,12 +13,16 @@ from flask_login import current_user, login_required
 from io import BytesIO
 from werkzeug.utils import secure_filename
 from codeathon import db, bcrypt
-from codeathon.models import Contest, Submission, Role, Team, User
+from codeathon.models import Contest, Language, Team, User
 from codeathon.admin.forms import (
     AddUserForm,
     AdminUpdateUserForm,
     ContestForm,
     ContestFormUpdate,
+    TeamForm,
+    TeamFormUpdate,
+    LanguageForm,
+    LanguageFormUpdate,
 )
 
 
@@ -41,7 +45,7 @@ def new_contest():
         db.session.add(contest)
         db.session.commit()
         flash("Your contest has been created!", "success")
-        return redirect(url_for("admin.contest", contest_id=contest.id))
+        return redirect(url_for("admin.contests_table"))
     return render_template(
         "admin/contest_add.html",
         title="New Contest",
@@ -63,7 +67,7 @@ def contests_table():
         abort(403)
     contests = Contest.query
     return render_template(
-        "admin/contests_table.html", title="Contests Table", contests=contests
+        "admin/contests_table.html", title="Contests", contests=contests
     )
 
 
@@ -79,7 +83,6 @@ def update_contest(contest_id):
         contest.description = form.description.data
         contest.start_date_time = form.start_date_time.data
         contest.end_date_time = form.end_date_time.data
-
         db.session.commit()
         flash("Your contest has been updated!", "success")
         return redirect(url_for("admin.contest", contest_id=contest.id))
@@ -102,6 +105,13 @@ def delete_contest(contest_id):
     if current_user.user_role.id != 3:
         abort(403)
     contest = Contest.query.get_or_404(contest_id)
+    team = Team.query.filter_by(contest_id=contest.id).first()
+    if team:
+        flash(
+            "This contest is associated with a team and can not be deleted until the associated team is deleted.",
+            "warning",
+        )
+        return redirect(url_for("admin.contest", contest_id=contest.id))
     db.session.delete(contest)
     db.session.commit()
     flash("Your contest has been deleted!", "success")
@@ -111,7 +121,151 @@ def delete_contest(contest_id):
 #######Languages#######
 
 
+@admin.route("/languages_table")
+def languages_table():
+    if current_user.user_role.id != 3:
+        abort(403)
+    languages = Language.query
+    return render_template(
+        "admin/languages_table.html", title="Programming Languages", languages=languages
+    )
+
+
+@admin.route("/language_add", methods=["GET", "POST"])
+def language_add():
+    if current_user.user_role.id != 3:
+        abort(403)
+    form = LanguageForm()
+    if form.validate_on_submit():
+        language = Language(
+            name=form.name.data,
+        )
+        db.session.add(language)
+        db.session.commit()
+        flash("The language has been created!", "success")
+        return redirect(url_for("admin.languages_table"))
+    return render_template(
+        "admin/language_add.html", title="languages", form=form, legend="Add language"
+    )
+
+
+@admin.route("/language/<int:language_id>/delete", methods=["POST"])
+@login_required
+def delete_language(language_id):
+    if current_user.user_role.id != 3:
+        abort(403)
+    language = Language.query.get_or_404(language_id)
+    db.session.delete(language)
+    db.session.commit()
+    flash("The language has been deleted!", "success")
+    return redirect(url_for("admin.language_table"))
+
+
+@admin.route("/language/<int:language_id>/update", methods=["GET", "POST"])
+@login_required
+def update_language(language_id):
+    if current_user.user_role.id != 3:
+        abort(403)
+    language = Language.query.get_or_404(language_id)
+    form = LanguageFormUpdate()
+    if form.validate_on_submit():
+        language.name = form.name.data
+        db.session.commit()
+        flash("Your language has been updated!", "success")
+        return redirect(url_for("admin.languages_table"))
+    elif request.method == "GET":
+        form.name.data = language.name
+    return render_template(
+        "admin/language_add.html",
+        title="Update Language",
+        form=form,
+        legend="Update Language",
+    )
+
+
 #######Teams#######
+@admin.route("/team/<int:team_id>")
+@login_required
+def team(team_id):
+    team = Team.query.get_or_404(team_id)
+    contests = Contest.query.order_by(Contest.start_date_time.desc())
+    return render_template(
+        "admin/team.html", title=team.name, team=team, contests=contests
+    )
+
+
+@admin.route("/teams_table")
+def teams_table():
+    if current_user.user_role.id != 3:
+        abort(403)
+    teams = Team.query
+    return render_template("admin/teams_table.html", title="Teams", teams=teams)
+
+
+@admin.route("/team_add", methods=["GET", "POST"])
+def team_add():
+    if current_user.user_role.id != 3:
+        abort(403)
+    form = TeamForm()
+    form.contest.choices = [
+        (contest.id, contest.title)
+        for contest in Contest.query.order_by(Contest.start_date_time.desc()).all()
+    ]
+    if form.validate_on_submit():
+        team = Team(
+            name=form.name.data,
+            contest_id=form.contest.data,
+        )
+        db.session.add(team)
+        db.session.commit()
+        flash("The team has been created!", "success")
+        return redirect(url_for("admin.teams_table"))
+    return render_template(
+        "admin/team_add.html",
+        title="Teams",
+        form=form,
+        legend="Add Team",
+    )
+
+
+@admin.route("/team/<int:team_id>/delete", methods=["POST"])
+@login_required
+def delete_team(team_id):
+    if current_user.user_role.id != 3:
+        abort(403)
+    team = Team.query.get_or_404(team_id)
+    db.session.delete(team)
+    db.session.commit()
+    flash("The team has been deleted!", "success")
+    return redirect(url_for("admin.team_table"))
+
+
+@admin.route("/team/<int:team_id>/update", methods=["GET", "POST"])
+@login_required
+def update_team(team_id):
+    if current_user.user_role.id != 3:
+        abort(403)
+    team = Team.query.get_or_404(team_id)
+    form = TeamFormUpdate()
+    form.contest.choices = [
+        (contest.id, contest.title)
+        for contest in Contest.query.order_by(Contest.start_date_time.desc()).all()
+    ]
+    if form.validate_on_submit():
+        team.name = form.name.data
+        team.contest_id = form.contest.data
+        db.session.commit()
+        flash("Your team has been updated!", "success")
+        return redirect(url_for("admin.teams_table"))
+    elif request.method == "GET":
+        form.name.data = team.name
+        form.contest.data = team.contest_id
+    return render_template(
+        "admin/team_add.html",
+        title="Update team",
+        form=form,
+        legend="Update team",
+    )
 
 
 #######Users#######
@@ -120,7 +274,7 @@ def users_table():
     if current_user.user_role.id != 3:
         abort(403)
     users = User.query
-    return render_template("admin/users_table.html", title="Users Table", users=users)
+    return render_template("admin/users_table.html", title="Users", users=users)
 
 
 @admin.route("/user_add", methods=["GET", "POST"])
